% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-cheers.R
\name{roclet_output.roclet_cheers}
\alias{roclet_output.roclet_cheers}
\title{roclet_output roclet_cheers method}
\usage{
roclet_output.roclet_cheers(x, results, base_path, ...)
}
\arguments{
\item{...}{}
}
\description{
roclet_output roclet_cheers method
}
