% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-cheers.R
\name{roclet_process.roclet_cheers}
\alias{roclet_process.roclet_cheers}
\title{roclet_process roclet_cheers method}
\usage{
roclet_process.roclet_cheers(x, blocks, env, base_path)
}
\arguments{
\item{base_path}{}
}
\description{
roclet_process roclet_cheers method
}
